Allow separate testers for secret readers and writers
